triplanar mapping with normals and ORDp

+++ Source/MurderInSpace/Public/Actors/IHasMesh.h
@@ -14,7 +14,7 @@ class IHasMesh
     
 public:
     virtual UPrimitiveComponent* GetMesh() const = 0;
-    virtual float GetMyMass() const = 0;
+    virtual double GetMyMass() const = 0;
     virtual FVector GetMyInertiaTensor() const { return FVector(1., 1., 1.); }
 };
