+++ Source/MurderInSpace/Public/Actors/MyPawn_SkeletalMesh.h
@@ -18,7 +18,7 @@ public:
 	AMyPawn_SkeletalMesh();
 
 	virtual UPrimitiveComponent* GetMesh() const override final { return SkeletalMesh; }
-	virtual float GetMyMass() const override { return pow(SkeletalMesh->Bounds.SphereRadius, 3); }
+	virtual double GetMyMass() const override { return pow(SkeletalMesh->Bounds.SphereRadius, 3); }
 	
 protected:
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly)
